requirements.txt
fastapi==0.111.0
uvicorn==0.30.1

sqlalchemy==2.0.30
pymysql==1.1.1

python-jose==3.3.0
passlib[bcrypt]==1.7.4

python-multipart==0.0.9

pydantic==2.7.4

argon2-cffi
